SurfacePosition
Initial attempt at implementing SurfacePosition. Algorithm is mostly shared with AirbornePosition, so the common elements are moved to the PositionUpdate class, which maintains a cache of even/odd frames, decides whether Global or Local calculation can be used, and generally contains the complexity of the CPR system. Some further sanity checks are required, and also a way to choose between multiple possible solutions for lat/lon for Surface data.